requirements.txt
numpy~=1.26.4
requests~=2.32.5
gradio~=5.47.2
transformers[torch]~=4.53.0
pillow~=11.3.0
einops~=0.8.1
torchvision~=0.23.0
tensorflow~=2.20.0
